% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active-clubs.R
\name{active_clubs_times}
\alias{active_clubs_times}
\title{Load Active Book Club Times}
\usage{
active_clubs_times(refresh = FALSE)
}
\arguments{
\item{refresh}{Set this to \code{TRUE} to check for updates to the Google Sheet.}
}
\value{
A tibble with columns \code{club}, \code{day_utc}, \code{hour_utc}, \code{facilitator},
and \code{book_abbr}.
}
\description{
Active book club times are stored in a Google Sheet, then parsed into an RDS.
This function loads the RDS if it is up-to-date, or updates the RDS if it is
out of date. Once this function has been called in a session, it always
returns the same result unless it is told to refresh.
}
