% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approved-books.R
\name{approved_books}
\alias{approved_books}
\title{Load Approved Books}
\usage{
approved_books(refresh = FALSE)
}
\arguments{
\item{refresh}{Set this to \code{TRUE} to check for updates to the Google Sheet.}
}
\value{
A tibble with columns TBD.
}
\description{
Approved books are stored in a Google Sheet, then parsed into an RDS.
This function loads the RDS if it is up-to-date, or updates the RDS if it is
out of date. Once this function has been called in a session, it always
returns the same result unless it is told to refresh.
}
