% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookclubdata-package.R
\docType{package}
\name{bookclubdata-package}
\alias{bookclubdata}
\alias{bookclubdata-package}
\title{bookclubdata: Process R4DS Book Club Data}
\description{
Functions to fetch, parse, and save R4DS book club data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/bookclubdata}
  \item \url{https://r4ds.github.io/bookclubdata/}
  \item Report bugs at \url{https://github.com/r4ds/bookclubdata/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
