% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilitators.R
\name{facilitator_volunteers_read}
\alias{facilitator_volunteers_read}
\title{Load new facilitator volunteers}
\usage{
facilitator_volunteers_read(refresh = FALSE)
}
\arguments{
\item{refresh}{Set this to \code{TRUE} to check for updates to the Google Sheet.}
}
\value{
A tibble with columns \code{facilitator_id}, \code{facilitator_handle}, and
\code{book_abbr}.
}
\description{
Facilitator volunteers are loaded into a Google Sheet (eventually via a form,
perhaps), then parsed into an RDS. This function loads the RDS if it is
up-to-date, or updates the RDS if it is out of date. Once this function has
been called in a session, it always returns the same result unless it is told
to refresh.
}
