% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_datetimes_utc}
\alias{make_datetimes_utc}
\title{Make datetimes next week}
\usage{
make_datetimes_utc(days, hours, timezones)
}
\arguments{
\item{days}{A vector of English day names ("Monday", "Tuesday", etc).}

\item{hours}{A vector of hours.}

\item{timezones}{A vector of starting timezones.}
}
\value{
A vector of equivalent UTC times "next week."
}
\description{
Convert a vector of days, hours, and timezones to datetimes "next week" in
UTC. The dates are at least 7 days in the future.
}
\examples{
make_datetimes_utc(
  days = c("Wednesday", "Saturday", "Thursday", "Monday"),
  hours = c(12L, 2L, 14L, 15L),
  timezones = c(
    "America/Chicago", "Europe/Rome", "Asia/Calcutta", "America/New_York"
  )
)
}
