% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signups.R
\name{signups}
\alias{signups}
\title{Process Signups}
\usage{
signups(refresh = FALSE)
}
\arguments{
\item{refresh}{Set this to \code{TRUE} to check for updates to the Google Sheet.}
}
\value{
A tibble with columns \code{user_name}, \code{user_id}, \code{book_name},
\code{timezone}, and \code{datetime_utc}.
}
\description{
Book club signups are logged in a Google Sheet, then parsed into an RDS. This
function loads the RDS if it is up-to-date, or updates the RDS if it is out
of date. Once this function has been called in a session, it always returns
the same result unless it is told to refresh.
}
