% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per-book.R
\name{signups_clear_user_book}
\alias{signups_clear_user_book}
\title{Remove book signups for a user}
\usage{
signups_clear_user_book(user_id, book_name)
}
\arguments{
\item{user_id}{A string identifying the user.}

\item{book_name}{The name of the book to read signups for.}
}
\value{
The signups tibble for that book.
}
\description{
Clear all signups for a given user and book, and resave the associated rds.
}
