% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per-book.R
\name{signups_read}
\alias{signups_read}
\title{Read the signups rds for a book}
\usage{
signups_read(book_name, refresh = FALSE)
}
\arguments{
\item{book_name}{The name of the book to read signups for.}

\item{refresh}{Whether to check for an update to the signups.}
}
\value{
A tibble with columns \code{user_name}, \code{user_id}, \code{timezone}, and
\code{datetime_utc}.
}
\description{
Read signups rds for a particular book.
}
