% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per-book.R
\name{signups_write}
\alias{signups_write}
\title{Write the signups rds for a book}
\usage{
signups_write(user_times_df, book_name)
}
\arguments{
\item{user_times_df}{A tibble with columns \code{user_name}, \code{user_id},
\code{timezone}, and \code{datetime_utc}.}

\item{book_name}{The name of the book to read signups for.}
}
\value{
The signups tibble for that book.
}
\description{
Update or create the signups rds for a particular book.
}
