% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{tz_minutes}
\alias{tz_minutes}
\title{Convert datetime minutes to 0 or 30}
\usage{
tz_minutes(timezones, datetimes = lubridate::now())
}
\arguments{
\item{timezones}{A character vector of timezones.}

\item{datetimes}{An optional vector of datetimes.}
}
\value{
A vector of 0L or 30L for each timezone.
}
\description{
Determine the minute offset of timezones from UTC on particular dates.
}
\examples{
tz_minutes(c("America/Chicago", "Europe/Rome", "Asia/Calcutta"))
}
